Fix: No NoSuchFileException thrown when unlocking a file twice

+++ src/main/java/applist/LockDownloadFile.java
@@ -9,9 +9,9 @@ package applist;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -21,14 +21,17 @@ package applist;
  */
 
 
+import com.github.vatbub.common.core.logging.FOKLogger;
 import com.github.vatbub.common.updater.Version;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.charset.Charset;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
+import java.nio.file.NoSuchFileException;
+import java.util.logging.Level;
 
 public class LockDownloadFile {
     private App app;
@@ -50,11 +53,16 @@ public class LockDownloadFile {
     }
 
     public void lock() throws IOException {
-        FileUtils.writeStringToFile(getLockFile(), "locked", Charset.forName("UTF-8"));
+        FileUtils.writeStringToFile(getLockFile(), "locked", StandardCharsets.UTF_8);
     }
 
     public void unlock() throws IOException {
-        Files.delete(getLockFile().toPath());
+        if (!isLocked()) return;
+        try {
+            Files.delete(getLockFile().toPath());
+        } catch (NoSuchFileException e) {
+            FOKLogger.log(getClass().getName(), Level.WARNING, "Unable to delete the lock file due to an exception", e);
+        }
     }
 
     public boolean isLocked() {
